pyproject.toml
[build-system]
requires = ["setuptools>=68", "build", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentbeats"
# Version set to initial; this will be updated by semantic-release
version = "1.2.5"
description = "AgentBeats SDK for building security-contest AI agents"
requires-python = ">=3.11"
authors = [{ "name" = "Simon Xie", "email" = "simonxie2004@outlook.com" }]
dependencies = [
    "tomli>=2",
    "toml",
    "fastmcp",
    "fastapi",
    "uvicorn",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.2.1",
    "sqlmodel>=0.0.21",
    "google-cloud-run>=0.10.10",
    "docker>=7.1.0",
    "httpx",
    "a2a-sdk>=0.3.0",
    "paramiko>=3.5.1",
    "openai-agents",
    "openai",
    "sqlalchemy",
    "supabase",
    "PyJWT",
    "python-dotenv",
    "python-multipart",
    "google-cloud-storage",
    "typer>=0.20.0",
    "pandas>=2.3.3",
    "litellm>=1.80.9",
    "pyarrow>=22.0.0",
    "fastparquet>=2024.11.0",
]
license = { text = "MIT" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]

[project.scripts]
agentbeats = "agentbeats.cli:typer_app"
ab = "agentbeats.cli:typer_app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
backend = ["**/*"]
mcpcp = ["**/*"]
agentbeats = ["frontend/*.html"]

# To ensure CI tools can find the version
[tool.semantic_release]
version_source   = "pyproject.toml"
version_variable = "project.version"
tag_format       = "v{version}"
repository_url   = "https://github.com/agentbeats/agentbeats.git"
changelog_file   = false
version_toml     = ["pyproject.toml:project.version"]
